requirements.txt
google.generativeai
Flask
requests
jsonify
python-dotenv
gunicorn
